pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "gemini-code-review"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pydantic>=2.11.7",
    "pydantic-ai>=1.0.5",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "tiktoken>=0.11.0",
]

[project.scripts]
gemini_code_review = "gemini_code_review.cli:main"
